fix(seeder): move phenomenon seeding into deploy job so production gets seeded
PhenomenonSeeder lived in EcoData.Sensors.Ingestion, which has no
PublishAsAzureContainerApp call in AppHost.cs and therefore never runs in
production. Phenomena and USGS parameter mappings were silently absent.

Move PhenomenonCatalog, UsgsParameterMappings, and the seeding logic into
EcoData.Seeder (which is published as an Azure Container App Job and runs
on every deploy). The two seed routines fit alongside the existing seeds
in DatabaseSeederWorker and reuse the already-referenced DbContexts. The
USGS data source lookup hits OrganizationDbContext.DataSources directly
instead of pulling in IDataSourceRepository.

Ingestion keeps ReadingBackfillService and UsgsIngestionWorker; only the
phenomenon seeder registration is removed.
