+++ src/Host/EcoData.Seeder/PhenomenaSeedData.cs
@@ -1,135 +1,6 @@
-using EcoData.Organization.DataAccess.Interfaces;
-using EcoData.Sensors.Database;
 using EcoData.Sensors.Database.Models;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 
-namespace EcoData.Sensors.Ingestion.Seeders;
-
-public sealed class PhenomenonSeeder(
-    IDbContextFactory<SensorsDbContext> contextFactory,
-    IDataSourceRepository dataSourceRepository,
-    ILogger<PhenomenonSeeder> logger
-) : IHostedService
-{
-    private const string UsgsDataSourceName = "USGS Puerto Rico";
-
-    public async Task StartAsync(CancellationToken cancellationToken)
-    {
-        await SeedPhenomenaAsync(cancellationToken);
-        await SeedUsgsParameterMappingsAsync(cancellationToken);
-    }
-
-    public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
-
-    private async Task SeedPhenomenaAsync(CancellationToken ct)
-    {
-        await using var context = await contextFactory.CreateDbContextAsync(ct);
-
-        var existingCodes = await context
-            .Phenomena.Select(p => p.Code)
-            .ToListAsync(ct);
-
-        var existingSet = existingCodes.ToHashSet(StringComparer.OrdinalIgnoreCase);
-        var now = DateTimeOffset.UtcNow;
-
-        var toAdd = PhenomenonCatalog
-            .All.Where(p => !existingSet.Contains(p.Code))
-            .Select(p => new Phenomenon
-            {
-                Id = Guid.CreateVersion7(),
-                Code = p.Code,
-                Name = p.Name,
-                Description = p.Description,
-                CanonicalUnit = p.CanonicalUnit,
-                DefaultValueShape = p.DefaultValueShape,
-                Capabilities = p.Capabilities,
-                CreatedAt = now,
-            })
-            .ToList();
-
-        if (toAdd.Count == 0)
-        {
-            logger.LogDebug("All {Total} phenomena already seeded", PhenomenonCatalog.All.Count);
-            return;
-        }
-
-        context.Phenomena.AddRange(toAdd);
-        await context.SaveChangesAsync(ct);
-        logger.LogInformation("Seeded {Count} phenomena", toAdd.Count);
-    }
-
-    private async Task SeedUsgsParameterMappingsAsync(CancellationToken ct)
-    {
-        var dataSource = await dataSourceRepository.GetByNameAsync(UsgsDataSourceName, ct);
-        if (dataSource is null)
-        {
-            logger.LogWarning(
-                "USGS data source '{Name}' not found; skipping USGS parameter mapping seed. Mappings will be seeded on next worker run after the data source is created.",
-                UsgsDataSourceName
-            );
-            return;
-        }
-
-        await using var context = await contextFactory.CreateDbContextAsync(ct);
-
-        var phenomenaByCode = await context
-            .Phenomena.ToDictionaryAsync(p => p.Code, p => p.Id, StringComparer.OrdinalIgnoreCase, ct);
-
-        var existingCodes = await context
-            .Parameters.Where(p => p.SourceId == dataSource.Id)
-            .Select(p => p.Code)
-            .ToListAsync(ct);
-
-        var existingSet = existingCodes.ToHashSet(StringComparer.OrdinalIgnoreCase);
-        var now = DateTimeOffset.UtcNow;
-
-        var toAdd = new List<Parameter>();
-        foreach (var mapping in UsgsParameterMappings.All)
-        {
-            if (existingSet.Contains(mapping.Code))
-            {
-                continue;
-            }
-            if (!phenomenaByCode.TryGetValue(mapping.PhenomenonCode, out var phenomenonId))
-            {
-                logger.LogWarning(
-                    "USGS code {Code} maps to phenomenon '{PhenomenonCode}' which is not in the catalog; skipping",
-                    mapping.Code,
-                    mapping.PhenomenonCode
-                );
-                continue;
-            }
-
-            toAdd.Add(new Parameter
-            {
-                Id = Guid.CreateVersion7(),
-                SourceId = dataSource.Id,
-                Code = mapping.Code,
-                Name = mapping.Name,
-                DefaultUnit = mapping.SourceUnit,
-                SensorTypeId = null,
-                PhenomenonId = phenomenonId,
-                SourceUnit = mapping.SourceUnit,
-                UnitFactor = mapping.UnitFactor,
-                UnitOffset = mapping.UnitOffset,
-                ValueShape = mapping.ValueShape,
-                CreatedAt = now,
-            });
-        }
-
-        if (toAdd.Count == 0)
-        {
-            logger.LogDebug("All {Total} USGS parameter mappings already seeded", UsgsParameterMappings.All.Count);
-            return;
-        }
-
-        context.Parameters.AddRange(toAdd);
-        await context.SaveChangesAsync(ct);
-        logger.LogInformation("Seeded {Count} USGS parameter mappings", toAdd.Count);
-    }
-}
+namespace EcoData.Seeder;
 
 internal sealed record PhenomenonSeed(
     string Code,
